reorganize for single command
