--- tools/monorepo-utils/src/changelog/commands/index.ts
@@ -1,15 +1,0 @@
-/**
- * External dependencies
- */
-import { Command } from '@commander-js/extra-typings';
-
-/**
- * Internal dependencies
- */
-import { extractCommand } from './extract';
-
-const program = new Command( 'changelog' )
-	.description( 'Changelog utilities' )
-	.addCommand( extractCommand );
-
-export default program;
