+++ tools/monorepo-utils/src/index.ts
@@ -10,7 +10,7 @@ import chalk from 'chalk';
  */
 import CodeFreeze from './code-freeze/commands';
 import Slack from './slack/commands/slack';
-import Changelog from './changelog/commands';
+import Changelog from './changelog';
 import { Logger } from './core/logger';
 import { isGithubCI } from './core/environment';
